Namespace function to_yaml()
