+++ lib/puppet/functions/to_yaml.rb
@@ -1,32 +1,14 @@
 # frozen_string_literal: true
 
-require 'yaml'
-# @summary
-#   Convert a data structure and output it as YAML
-Puppet::Functions.create_function(:to_yaml) do
-  # @param data
-  #   The data you want to convert to YAML
-  # @param options
-  #   A hash of options that will be passed to Ruby's Psych library. Note, this could change between Puppet versions, but at time of writing these are `line_width`, `indentation`, and `canonical`.
-  #
-  # @example Output YAML to a file
-  #   file { '/tmp/my.yaml':
-  #     ensure  => file,
-  #     content => to_yaml($myhash),
-  #   }
-  # @example Use options to control the output format
-  #   file { '/tmp/my.yaml':
-  #     ensure  => file,
-  #     content => to_yaml($myhash, {indentation => 4})
-  #   }
-  #
-  # @return [String] The YAML document
-  dispatch :to_yaml do
-    param 'Any', :data
-    optional_param 'Hash', :options
-  end
+# THIS FILE WAS GENERATED BY `rake regenerate_unamespaced_shims`
 
-  def to_yaml(data, options = {})
-    data.to_yaml(options.transform_keys(&:to_sym))
+# @summary DEPRECATED.  Use the namespaced function [`stdlib::to_yaml`](#stdlibto_yaml) instead.
+Puppet::Functions.create_function(:to_yaml) do
+  dispatch :deprecation_gen do
+    repeated_param 'Any', :args
+  end
+  def deprecation_gen(*args)
+    call_function('deprecation', 'to_yaml', 'This function is deprecated, please use stdlib::to_yaml instead.')
+    call_function('stdlib::to_yaml', *args)
   end
 end
