+++ spec/functions/to_yaml_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe 'to_yaml' do
+describe 'stdlib::to_yaml' do
   it { is_expected.not_to be_nil }
   it { is_expected.to run.with_params('').and_return("--- ''\n") }
   it { is_expected.to run.with_params(true).and_return(%r{--- true\n}) }
